feat: :sparkles: feat: 루트 컴포넌트 추가, 소켓 훅 호출부 해당 컴포넌트로 옮김, 에러 바운더리 추가
예외처리는 각 컴포넌트에서 진행하지 않고 에러 바운더리로 위임하는 것으로 하겠슴니다
