+++ frontend/src/common/index.ts
@@ -3,3 +3,4 @@ export { LoginLayout } from './LoginLayout';
 export { Modal } from './Modal';
 export { Dropdown } from './Dropdown';
 export { Tooltip } from './Tooltip';
+export { RootComponent } from './RootComponent';
